requirements.txt
# Requirements for 
# To install everything, run pip install -r requirements.txt

pyserial        # needed for serial communication with arduino
numpy           # used to process large sets of data
PyQt6           # needed for UI
pyqtgraph       # needed for graphing